adding push command
