+++ src/common/push.ui.ts
@@ -1,10 +1,5 @@
 import {Connection} from '@salesforce/core';
 
-interface UiReturn {
-  uiRecords: string[];
-  uiPmsToUpload: Set<string>;
-}
-
 export interface PushPmlParams {
   sourcepath: string;
   conn: Connection;
@@ -12,15 +7,12 @@ export interface PushPmlParams {
   uisToUpload: Set<string>;
 }
 
-export async function pushUI(params: PushPmlParams): Promise<UiReturn> {
+export async function pushUI(params: PushPmlParams): Promise<string[]> {
   const { sourcepath, conn, pushAll, uisToUpload } = params;
   console.log(sourcepath)
   console.log(conn.toString())
   console.log(pushAll.toString())
   console.log(uisToUpload.toString())
   await Promise.resolve(); // todo remove, added for eslint error
-  return {
-    uiRecords: [],
-    uiPmsToUpload: new Set<string>()
-  }
+  return []
 }
